Rename kotlin_native -> kotlin.
