+++ backend.native/tests/runtime/basic/hello0.kt
@@ -1,4 +1,4 @@
 // TODO: remove kotlin_native.io once overrides are in place.
 fun main(args : Array<String>) {
-  kotlin_native.io.print("Hello, world!")
+  kotlin.io.print("Hello, world!")
 }
