+++ backend.native/tests/runtime/basic/hello1.kt
@@ -1,4 +1,4 @@
 // TODO: remove kotlin_native.io once overrides are in place.
 fun main(args : Array<String>) {
-  kotlin_native.io.print(kotlin_native.io.readLine())
+  kotlin.io.print(kotlin.io.readLine())
 }
