+++ runtime/src/main/kotlin/kotlin/Any.kt
@@ -1,14 +1,14 @@
-package kotlin_native
+package kotlin
 
 @ExportTypeInfo("theAnyTypeInfo")
-public open class Any0 {
+public open class Any {
     // @SymbolName("Kotlin_Any_equals")
     // external public open operator fun equals0(other: Any0?): Boolean
 
     // TODO: do we need that in Any?
     @SymbolName("Kotlin_Any_hashCode")
-    external public open fun hashCode0(): Int
+    external public open fun hashCode(): Int
 
     @SymbolName("Kotlin_Any_toString")
-    external public open fun toString0(): String
+    external public open fun toString(): String
 }
