+++ runtime/src/main/kotlin/kotlin/String.kt
@@ -1,4 +1,4 @@
-package kotlin_native
+package kotlin
 
 @SymbolName("Kotlin_String_fromUtf8Array")
 external fun fromUtf8Array(array: ByteArray) : String
