+++ runtime/src/main/kotlin/kotlin/io/Console.kt
@@ -1,0 +1,7 @@
+package kotlin.io
+
+@kotlin.SymbolName("Kotlin_io_Console_print")
+external public fun print(message: kotlin.String)
+
+@kotlin.SymbolName("Kotlin_io_Console_readLine")
+external public fun readLine(): kotlin.String
